Update ResponseValidator.php
Issue #35
I debugged the behaviour, the response from transmission rpc was fine, it was sending result object as "duplicate torrent" and argument object was empty. 
the validator for addtorrent was parsing it as invalid because of absence of torrent-duplicate object inside arguments.
I have fixed it and requested a pull.